Compare Two Directories for Synchronization Status

This function compares two directories, typically referred to as 'left' and 'right', to determine their synchronization status at the file level. The primary goal is to identify the synchronization status of files present in both directories and those exclusive to either directory.

Usage

compare_directories(
  left_path,
  right_path,
  recurse = TRUE,
  by_date = TRUE,
  by_content = FALSE,
  verbose = FALSE,
  short_paths = FALSE
)

Arguments

left_path

Path to the left/first directory.

right_path

Path to the right/second directory.

recurse

If TRUE, fully recurses through subdirectories. If a positive integer, specifies the number of levels to recurse.

by_date

Logical. If TRUE (default), compares directories based on the modification date of common files.

by_content

Logical. If TRUE, compares directories based on the hashed content of common files. Default is FALSE

verbose

Logical. If TRUE display additional info on the comparison process. Default is FALSE

short_paths

Logical. If TRUE, remove whole path from output for cleaner display. Default is FALSE

Value

A list of class "syncdr_status" containing the following elements:

Non-common files: Paths and synchronization status of files exclusive to either directory.

Common files: Paths and synchronization status of files present in both directories.

Path of the left directory.

Path of the right directory.

Sync Status Types

The synchronization status is determined for files present in both directories, as well as for files exclusive to either directory. It can be computed based on modification date only, content only, or both.

For Common Files:

When comparing by date: 'new', 'old', or 'same'.

When comparing by date and content: 'new and different', 'new and same', 'old and different', 'old and same', 'same and different', or 'same and same'.

When comparing by content only: 'different' or 'same'.

For Non-Common Files:

When comparing by date (or by date and content, or by content only): 'only in left' or 'only in right'.

Examples

# Compare directories with 'compare_directories()'
e <- toy_dirs()
#> ■■■■■■■■■■■■■■■                   47% | ETA:  5s
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■       87% | ETA:  1s

# Get left and right directories' paths
left  <- e$left
right <- e$right
compare_directories(left, right)
#> $common_files
#>                                                                   path_left
#> 1 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/B/B1.Rds
#> 2 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/B/B2.Rds
#> 3 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/C/C1.Rds
#> 4 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/C/C2.Rds
#> 5 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/C/C3.Rds
#> 6 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/D/D1.Rds
#> 7 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/D/D2.Rds
#>                                                                   path_right
#> 1 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/right/B/B1.Rds
#> 2 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/right/B/B2.Rds
#> 3 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/right/C/C1.Rds
#> 4 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/right/C/C2.Rds
#> 5 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/right/C/C3.Rds
#> 6 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/right/D/D1.Rds
#> 7 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/right/D/D2.Rds
#>   modification_time_left modification_time_right is_new_left is_new_right
#> 1    2024-07-05 14:44:00     2024-07-05 14:44:01       FALSE         TRUE
#> 2    2024-07-05 14:44:03     2024-07-05 14:44:04       FALSE         TRUE
#> 3    2024-07-05 14:44:01     2024-07-05 14:44:07       FALSE         TRUE
#> 4    2024-07-05 14:44:04     2024-07-05 14:44:05       FALSE         TRUE
#> 5    2024-07-05 14:44:06     2024-07-05 14:44:07       FALSE         TRUE
#> 6    2024-07-05 14:44:03     2024-07-05 14:44:02        TRUE        FALSE
#> 7    2024-07-05 14:44:06     2024-07-05 14:44:05        TRUE        FALSE
#>                         sync_status
#> 1 older in left, newer in right dir
#> 2 older in left, newer in right dir
#> 3 older in left, newer in right dir
#> 4 older in left, newer in right dir
#> 5 older in left, newer in right dir
#> 6 newer in left, older in right dir
#> 7 newer in left, older in right dir
#> 
#> $non_common_files
#> # A tibble: 9 × 3
#>   path_left                                               path_right sync_status
#> * <fs::path>                                              <fs::path> <chr>      
#> 1 …0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/A/A1.Rds NA         only in le…
#> 2 …0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/A/A2.Rds NA         only in le…
#> 3 …0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/A/A3.Rds NA         only in le…
#> 4 …0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/B/B3.Rds NA         only in le…
#> 5 NA                                                      …icate.Rds only in ri…
#> 6 NA                                                      …/D/D3.Rds only in ri…
#> 7 NA                                                      …/E/E1.Rds only in ri…
#> 8 NA                                                      …/E/E2.Rds only in ri…
#> 9 NA                                                      …/E/E3.Rds only in ri…
#> 
#> $left_path
#> /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left
#> 
#> $right_path
#> /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/right
#> 
#> attr(,"class")
#> [1] "syncdr_status"
# Compare by date and content
compare_directories(left, right, by_content = TRUE)
#> $common_files
#>                                                                   path_left
#> 1 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/B/B1.Rds
#> 2 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/B/B2.Rds
#> 3 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/C/C1.Rds
#> 4 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/C/C2.Rds
#> 5 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/C/C3.Rds
#> 6 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/D/D1.Rds
#> 7 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/D/D2.Rds
#>                                                                   path_right
#> 1 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/right/B/B1.Rds
#> 2 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/right/B/B2.Rds
#> 3 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/right/C/C1.Rds
#> 4 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/right/C/C2.Rds
#> 5 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/right/C/C3.Rds
#> 6 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/right/D/D1.Rds
#> 7 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/right/D/D2.Rds
#>   modification_time_left modification_time_right is_new_left is_new_right
#> 1    2024-07-05 14:44:00     2024-07-05 14:44:01       FALSE         TRUE
#> 2    2024-07-05 14:44:03     2024-07-05 14:44:04       FALSE         TRUE
#> 3    2024-07-05 14:44:01     2024-07-05 14:44:07       FALSE         TRUE
#> 4    2024-07-05 14:44:04     2024-07-05 14:44:05       FALSE         TRUE
#> 5    2024-07-05 14:44:06     2024-07-05 14:44:07       FALSE         TRUE
#> 6    2024-07-05 14:44:03     2024-07-05 14:44:02        TRUE        FALSE
#> 7    2024-07-05 14:44:06     2024-07-05 14:44:05        TRUE        FALSE
#>         sync_status is_diff
#> 1 different content    TRUE
#> 2 different content    TRUE
#> 3      same content   FALSE
#> 4 different content    TRUE
#> 5 different content    TRUE
#> 6 different content    TRUE
#> 7 different content    TRUE
#> 
#> $non_common_files
#> # A tibble: 9 × 3
#>   path_left                                               path_right sync_status
#> * <fs::path>                                              <fs::path> <chr>      
#> 1 …0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/A/A1.Rds NA         only in le…
#> 2 …0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/A/A2.Rds NA         only in le…
#> 3 …0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/A/A3.Rds NA         only in le…
#> 4 …0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/B/B3.Rds NA         only in le…
#> 5 NA                                                      …icate.Rds only in ri…
#> 6 NA                                                      …/D/D3.Rds only in ri…
#> 7 NA                                                      …/E/E1.Rds only in ri…
#> 8 NA                                                      …/E/E2.Rds only in ri…
#> 9 NA                                                      …/E/E3.Rds only in ri…
#> 
#> $left_path
#> /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left
#> 
#> $right_path
#> /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/right
#> 
#> attr(,"class")
#> [1] "syncdr_status"
# Compare by content only
compare_directories(left, right, by_content = TRUE, by_date = FALSE)
#> $common_files
#>                                                                   path_left
#> 1 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/B/B1.Rds
#> 2 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/B/B2.Rds
#> 3 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/C/C1.Rds
#> 4 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/C/C2.Rds
#> 5 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/C/C3.Rds
#> 6 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/D/D1.Rds
#> 7 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/D/D2.Rds
#>                                                                   path_right
#> 1 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/right/B/B1.Rds
#> 2 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/right/B/B2.Rds
#> 3 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/right/C/C1.Rds
#> 4 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/right/C/C2.Rds
#> 5 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/right/C/C3.Rds
#> 6 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/right/D/D1.Rds
#> 7 /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/right/D/D2.Rds
#>   modification_time_left modification_time_right is_diff       sync_status
#> 1    2024-07-05 14:44:00     2024-07-05 14:44:01    TRUE different content
#> 2    2024-07-05 14:44:03     2024-07-05 14:44:04    TRUE different content
#> 3    2024-07-05 14:44:01     2024-07-05 14:44:07   FALSE      same content
#> 4    2024-07-05 14:44:04     2024-07-05 14:44:05    TRUE different content
#> 5    2024-07-05 14:44:06     2024-07-05 14:44:07    TRUE different content
#> 6    2024-07-05 14:44:03     2024-07-05 14:44:02    TRUE different content
#> 7    2024-07-05 14:44:06     2024-07-05 14:44:05    TRUE different content
#> 
#> $non_common_files
#> # A tibble: 9 × 3
#>   path_left                                               path_right sync_status
#> * <fs::path>                                              <fs::path> <chr>      
#> 1 …0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/A/A1.Rds NA         only in le…
#> 2 …0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/A/A2.Rds NA         only in le…
#> 3 …0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/A/A3.Rds NA         only in le…
#> 4 …0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left/B/B3.Rds NA         only in le…
#> 5 NA                                                      …icate.Rds only in ri…
#> 6 NA                                                      …/D/D3.Rds only in ri…
#> 7 NA                                                      …/E/E1.Rds only in ri…
#> 8 NA                                                      …/E/E2.Rds only in ri…
#> 9 NA                                                      …/E/E3.Rds only in ri…
#> 
#> $left_path
#> /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/left
#> 
#> $right_path
#> /var/folders/8y/hpp0r5m53kb_srnzjh912rmc0000gp/T/Rtmpov7fCm/right
#> 
#> attr(,"class")
#> [1] "syncdr_status"