Compare modification times of two files and determine synchronization status

This function compares the date of last modification of two files and determines their synchronization status

Usage

compare_modification_times(modification_time_left, modification_time_right)

Arguments

modification_time_left

modification time of the file in the left directory

modification_time_right

modification time of the file in the right directory

Value

A list containing the following components:

is_new_left

Logical. Indicates if the file in the left directory is newer (i.e., has a later modification time) than the file in the right directory

is_new_right

Logical. Indicates if the file in the right directory is newer (i.e., has a later modification time) than the file in the left directory

sync_status_date

Character. Describes the synchronization status between the two files based on their modification times: - "newer in left, older in right dir": Left file is newer than right file. - "older in left, newer in right dir": Right file is newer than left file. - "same date": Both files have the same modification time.