Create toy directories to test syncdr functions

create directories in syncdr environment. Directories are deleted when a new R session is started

Usage

toy_dirs(verbose = FALSE)

Arguments

verbose

logical: display information. Default is FALSE

Value

syncdr environment with toy directory paths, i.e., left and right paths

Details

This function is a little slow because it must use Sys.sleep() to save files with the same name but different time stamp.

Examples

toy_dirs(verbose = TRUE)
#> ■■■■■■■                           20% | ETA:  8s
#> /tmp/Rtmp3arur5/left
#> ├── A
#> │   ├── A1.Rds
#> │   ├── A2.Rds
#> │   └── A3.Rds
#> ├── B
#> │   ├── B1.Rds
#> │   ├── B2.Rds
#> │   └── B3.Rds
#> ├── C
#> │   ├── C1.Rds
#> │   ├── C1_duplicate.Rds
#> │   ├── C2.Rds
#> │   └── C3.Rds
#> ├── D
#> │   ├── D1.Rds
#> │   ├── D2.Rds
#> │   └── D3.Rds
#> └── E
#>     ├── E1.Rds
#>     ├── E2.Rds
#>     └── E3.Rds
#> /tmp/Rtmp3arur5/right
#> ├── A
#> │   ├── A1.Rds
#> │   ├── A2.Rds
#> │   └── A3.Rds
#> ├── B
#> │   ├── B1.Rds
#> │   ├── B2.Rds
#> │   └── B3.Rds
#> ├── C
#> │   ├── C1.Rds
#> │   ├── C1_duplicate.Rds
#> │   ├── C2.Rds
#> │   └── C3.Rds
#> ├── D
#> │   ├── D1.Rds
#> │   ├── D2.Rds
#> │   └── D3.Rds
#> └── E
#>     ├── E1.Rds
#>     ├── E2.Rds
#>     └── E3.Rds
#> ■■■■■■■■■■■■■■■■■                 53% | ETA:  6s